test (PlayerName) : String name 원시 값을 PlayerName으로 포장 테스트
